# chore(splunk_hec source): Fix handling of non UTF8 payloads