Closes #10778 

This PR handles non UTF8 payloads in incoming requests using `String::from_utf8_lossy`. This matches Splunk HEC's behavior of accepting such payloads and replacing invalid bytes with the replacement character (tested against `v8.2.4`).

As mentioned in the ticket, potential follow up work could include configuring this behavior and sending to a separate output for invalid payloads. Let me know if this change should be included in the `v0.20.0` upgrade guide.

<!--
**Your PR title must conform to the conventional commit spec!**

  <type>(<scope>)!: <description>

  * `type` = chore, enhancement, feat, fix, docs
  * `!` = OPTIONAL: signals a breaking change
  * `scope` = Optional when `type` is "chore" or "docs", available scopes https://github.com/timberio/vector/blob/master/.github/semantic.yml#L20
  * `description` = short description of the change

Examples:

  * enhancement(file source): Add `sort` option to sort discovered files
  * feat(new source): Initial `statsd` source
  * fix(file source): Fix a bug discovering new files
  * chore(external docs): Clarify `batch_size` option
-->