@001wwang 2022-01-19
> Nice! I am a bit wary that now the entire request body needs to be read into memory and decompressed before we extract events from it. I believe previously it was incrementally decoding them. Albeit, I'm guessing that reading it all at once is probably contributing to the performance gain here.
> 
> I'm curious if @binarylogic or @blt have any thoughts about what we should be preferring for sources. Incremental reading and streaming of events to keep memory consumption down at, probably, the risk of worse performance? Or reading in entire batches at once and decoding which will use more memory but, probably, perform faster?

Indeed, the overall performance increase here is interesting. @blt and I had discussed adding some micro-benchmarks around this code to get a better sense of whether this logic is faster or if it's creating some impact elsewhere. It would also be a good opportunity for me to learn `criterion` and such. Happy to dig into this more after wrapping up other tasks.