@blt 2022-01-19
The implementation of `from_utf8_lossy` is [here](https://doc.rust-lang.org/stable/src/alloc/string.rs.html#558). I can conceive of implementing something like this in terms of an iterator, but I'm not sure it's motivated.  You're right, @jszwedko, that this patch potentially uses more memory, but as I read it not terribly more theoretically than previous? The previous implementation would read the payload piecemeal and deserialize into a `JsonValue`, so we'd have to fit whatever the chunk size of the reader was in plus the `JsonValue` we built up. Now we fit the whole payload plus the `JsonValue`, yes? So maybe 2X whatever the eventual size of `JsonValue` is at peak usage versus the previous 1.xX usage?

Personally I'd vote we merge this as-is and if we can construct a payload that causes a serious memory blow-up then we address by making an iterator consuming `from_utf8_lossy` variant.