@jszwedko 2022-01-19
> To add context, throughput is king. So using more memory (within reason) for higher performance is preferable. Ideally users could opt out of this behavior, but that would be an advanced setting to reduce the batch sizes.

👍 this is a good overall rule-of-thumb.