@001wwang 2022-01-27
(Kudos to @blt for walking me through this!)
### Summary
At face value, the change introduced in this PR looks like it would cause a performance regression. Instead of simply lazily iterating through and deserializing the request body into `Event`s, we're now reading the entire request body upfront, replacing any invalid UTF-8 bytes, and then proceeding as before. 

Yet, the soak results show a (sizeable) performance _increase_ for `splunk_hec` related soaks. What?! Surely then, the new implementation is actually faster than before, and that's why performance increased? You might even begin to hypothesize some seriously convincing reasons why that could be.

Unfortunately, no. The new implementation is, like you initially suspected, slower than the original (see Details section). This PR is just a great example of how slowing down one piece of a complex system can result in an overall speed up (for reasons that are hard to pin down).

### Details
To prove that this new change is indeed slower than the original implementation, we can extract the core logic, create some microbenchmarks, and compare the two implementations. [Here is a branch including the simple new benchmarks](https://github.com/vectordotdev/vector/compare/splunk-lossy-decoding-change-benchmark).

The core logic we're comparing is as follows:
```rust
// new implementation
let mut data = Vec::new();
let body = if payload.gzip {
    MultiGzDecoder::new(payload.body.reader())
        .read_to_end(&mut data)
        .unwrap();
    String::from_utf8_lossy(data.as_slice())
} else {
    String::from_utf8_lossy(payload.body.as_ref())
};

Deserializer::from_str(&body).into_iter::<Value>();
```
```rust
// old implementation
let reader: Box<dyn Read + Send> = if payload.gzip {
    Box::new(MultiGzDecoder::new(payload.body.reader()))
} else {
    Box::new(payload.body.reader())
};

Deserializer::from_reader(reader).into_iter::<Value>();
```
The JSON body used for input is simply
```
"{ "message": "hello world" }"
```

Running them, we get the following results (output simplified for legibility). `new` is the implementation introduced in this PR and `old` is the original. To explore both branches of the core logic, `-gzip` indicates whether the payload is gzipped.
```shell
vector::sources::splunk_hec/new                                                                             
                        time:   [24.407 ns 24.482 ns 24.555 ns]
vector::sources::splunk_hec/new-gzip                                                                             
                        time:   [4.8594 us 4.8820 us 4.9057 us]
vector::sources::splunk_hec/old                                                                             
                        time:   [18.194 ns 18.235 ns 18.272 ns]
vector::sources::splunk_hec/old-gzip                                                                            
                        time:   [1.6066 us 1.6433 us 1.6816 us]
```
As you can see, even in this simple scenario, the new implementation is indeed slower.